+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `K8s.Client.Mint.HTTPAdapter` - Monitor caller and cleanup state upon `:DOWN`.
+
 ## [2.0.0-rc.3] - 2023-01-01
 
 ### Fixed